Fixed auto-adding things to static datamap from warning marker. This broke when I changed the path separator to a dot, as this code was assuming a slash.
Probably a better design here would be to separate the model and view here -- the path shown in the warning is part of the view, and should not also serve as the model. This implies we would store the path in the desired form (looks like this code wants an array) in the warning marker object. So we'd have to subclass the warning marker so it could hold this extra data.